Add order specific flowgraph function
The allBlocks flowgraph helper function will return a list of basic
blocks that are guaranteed to be in postorder, useful for bottom-up
traversing. If top-down traversing is desired, use a reverse
iterator on the list.

I have also added an initParents function, if the algorithm traversing
the flowgraph needs parent information, this function must be called
before calling allBlocks. Note that after any manipulation of the
flowgraph, parents are no longer guaranteed to be correct. Thus each
iteration through the flowgraph parents need to be reinitialized. Or
the algorithm must keep the parent information up to date during the
optimization iteration.

I have modified the MaximizeBlocks and register allocation to use
these helper functions.